software/backend/requirements.txt
pydantic~=2.12.4
fastapi~=0.121.0
python-dotenv~=0.21.0
psycopg~=3.2.12